build: Refactor file hirachy, introducing new makefiles
