+++ Sources/render/outputLine.cpp
