+++ Sources/render/renderer.h
@@ -3,8 +3,8 @@
 
 #include <vector>
 #include <SDL.h>
-#include "xbeMenuItem.h"
-#include "menuItem.h"
+#include "../menu/xbeMenuItem.h"
+#include "../menu/menuItem.h"
 
 int min(int lhs, int rhs);
 int max(int lhs, int rhs);
